Remove wfe1 integration tests
