--- test/config/wfe.json
@@ -1,58 +1,0 @@
-{
-  "wfe": {
-    "listenAddress": "0.0.0.0:4000",
-    "TLSListenAddress": "0.0.0.0:4430",
-    "serverCertificatePath": "test/wfe-tls/boulder/cert.pem",
-    "serverKeyPath": "test/wfe-tls/boulder/key.pem",
-    "allowOrigins": ["*"],
-    "shutdownStopTimeout": "10s",
-    "subscriberAgreementURL": "http://boulder:4000/terms/v1",
-    "debugAddr": ":8000",
-    "directoryCAAIdentity": "happy-hacker-ca.invalid",
-    "directoryWebsite": "https://github.com/letsencrypt/boulder",
-    "blockedKeyFile": "test/example-blocked-keys.yaml",
-    "tls": {
-      "caCertFile": "test/grpc-creds/minica.pem",
-      "certFile": "test/grpc-creds/wfe.boulder/cert.pem",
-      "keyFile": "test/grpc-creds/wfe.boulder/key.pem"
-    },
-    "raService": {
-      "serverAddress": "ra.boulder:9094",
-      "timeout": "20s"
-    },
-    "saService": {
-      "serverAddress": "sa.boulder:9095",
-      "timeout": "15s"
-    },
-    "getNonceService": {
-      "serverAddress": "nonce.boulder:9101",
-      "timeout": "15s"
-    },
-    "redeemNonceServices": {
-      "taro": {
-        "serverAddress": "nonce1.boulder:9101",
-        "timeout": "15s"
-      },
-      "zinc": {
-        "serverAddress": "nonce2.boulder:9101",
-        "timeout": "15s"
-      }
-    },
-    "features": {
-      "StripDefaultSchemePort": true
-    }
-  },
-
-  "syslog": {
-    "stdoutlevel": 4,
-    "sysloglevel": 6
-  },
-  "beeline": {
-      "mute": true,
-      "dataset": "Test"
-  },
-
-  "common": {
-    "issuerCert": "/hierarchy/intermediate-cert-rsa-a.pem"
-  }
-}
