+++ test/startservers.py
@@ -106,10 +106,6 @@ SERVICES = (
         4001, None,
         ('./bin/boulder-wfe2', '--config', os.path.join(config_dir, 'wfe2.json')),
         ('sd-test-srv', 'boulder-ra-1', 'boulder-ra-2', 'boulder-sa-1', 'boulder-sa-2', 'nonce-service-taro', 'nonce-service-zinc')),
-    Service('boulder-wfe',
-        4000, None,
-        ('./bin/boulder-wfe', '--config', os.path.join(config_dir, 'wfe.json')),
-        ('sd-test-srv', 'boulder-ra-1', 'boulder-ra-2', 'boulder-sa-1', 'boulder-sa-2', 'nonce-service-taro', 'nonce-service-zinc')),
     Service('log-validator',
         8016, None,
         ('./bin/log-validator', '--config', os.path.join(config_dir, 'log-validator.json')),
